requirements.txt
bokeh==0.12.3
dill==0.2.1
Flask==0.12
gunicorn==19.1.0
Jinja2==2.8
oauth==1.0.1
oauthlib==2.0.1
pandas==0.19.2
python-dotenv==0.6.1
requests>=2.12.4
simplejson==3.8.0
